Move utils to AudienceTile.
